remove intermediary client
